Added backwards compatibility for shift register lights
Added config.h which contains WS281x define, if this line is commented
out the old shift register lights implenetation will be used.
Also the LightsController class is now one class and set of files
instead of 2 different ones for shift register and WS2811 lights. This
will make it easier to fix bugs and implement new features in the
future.