Moved Angular WebUI to hardcoded Gzip strings
Instead of loading HTML , js and CSS files onto the SPIFFS, we now just included as an GZIP'ed string in the source code when compiling. This should allow for easier updates in the future.

Also added gulpfile to build all this stuff.

Also copied over latest from WebUI project

Thanks to Xose Perez of the Espurna project to introduce me to this method of building HTML files into the code